add:hook_myprofile
