+++ frontend/src/components/organisms/Footer.tsx
@@ -1,11 +1,11 @@
 import { FC } from 'react';
 import { Segment, Container } from 'semantic-ui-react';
-import Footercolumn from 'components/molecules/Footercolumn';
+import EnhancedFootercolumn from 'container/EnhancedFootercolumn';
 
 const Footer: FC = () => (
   <Segment inverted vertical style={{ padding: '4rem 0' }}>
     <Container>
-      <Footercolumn />
+      <EnhancedFootercolumn />
     </Container>
   </Segment>
 );
